⚙️ FEATURE-#201: Extract checkpoint to Flow base class and adapt strategies to use TaskEngine
