+++ dotflow/abc/flow.py
@@ -3,7 +3,9 @@
 from abc import ABC, abstractmethod
 from uuid import UUID
 
+from dotflow.core.context import Context
 from dotflow.core.task import Task
+from dotflow.core.types import TypeStatus
 
 
 class Flow(ABC):
@@ -40,3 +42,22 @@ class Flow(ABC):
     @abstractmethod
     def run(self) -> None:
         return None
+
+    def _has_checkpoint(self, task: Task) -> bool:
+        if not self.resume:
+            return False
+
+        context = task.config.storage.get(key=task.config.storage.key(task=task))
+
+        return context.storage is not None
+
+    def _restore_checkpoint(self, task: Task) -> Context:
+        previous_context = task.config.storage.get(
+            key=task.config.storage.key(task=task)
+        )
+
+        task.status = TypeStatus.COMPLETED
+        task.current_context = previous_context
+        self._flow_callback(task=task)
+
+        return previous_context
